Add Either monad
Represents a value of one of two possible types (a disjoint union.)
An instance of `Either` is either an instance of `Left` or `Right`.

A common use of `Either` is as an alternative to `Option` for dealing
with possible missing values.  In this usage, `None` is replaced
with a `Left` which can contain useful information.
`Right` takes the place of `Some`. Convention dictates
that `Left` is used for failure and `Right` is used for success.

For example, you could use `Either<String, Fixnum>` to detect whether a
received input is a `String` or an `Fixnum`.

```ruby
  in = Readline.readline('Type Either a string or an Int: ', true)
  result = begin
    Right(Integer(in))
  rescue ArgumentError
    Left(in)
  end

  puts(
    result.reduce(
      -> (x) { "You passed me the Int: #{x}, which I will increment. #{x} + 1 = #{x+1}" },
      -> (x) { "You passed me the String: #{x}" }
    )
  )
```

`Either` is right-biased, which means that `Right` is assumed to be the default case to
operate on. If it is `Left`, operations like `#map`, `#flat_map`, ... return the `Left` value
unchanged:

```ruby
  Right(12).map { |_| _ * 2) #=> Right(24)
  Left(23).map { |_| _ * 2)  #=> Left(23)
```

Inspired by original Scala's Either monad https://github.com/scala/scala/blob/2.12.x/src/library/scala/util/Either.scala